+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - fix and add doc.
 - add development dependency
 - avoid instance variable name conflict (`@models` to `@__models`)
+- add #empty?, #clear to InnerModelCollection
 
 ## [0.1.1] - 2024-01-08